requirements.txt
streamlit
pandas
spacy
neo4j
networkx
scikit-learn
numpy
scipy
googletrans
python-dotenv
huggingface_hub
sentence-transformers>=2.2.0
plotly
faiss-cpu
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
langchain-huggingface
langchain-core
transformers>=4.30.0